Move each tree item to it's own folder
