+++ lib/commands/refreshStateView.spec.js
@@ -1,6 +1,6 @@
 jest.mock('../helpers/pivotaly')
 const refreshStateView = require('./refreshStateView')
-const StateInfoDataProvider = require('../views/storyInfoDataProvider')
+const StateInfoDataProvider = require('../views/storyInfo/storyInfoDataProvider')
 const Context = require('../../test/factories/vscode/context')
 const ptlyHelpers = require('../helpers/pivotaly')
