+++ lib/commands/undeliverTask.spec.js
@@ -2,8 +2,8 @@ jest.mock('../validation/rebounds')
 jest.mock('../../model/tasks')
 const undeliverTask = require('./undeliverTask')
 const rebounds = require('../validation/rebounds')
-const TaskTreeItem = require('../views/taskAndBlockerTreeItem')
-const StoryInfoDataProvider = require('../views/storyInfoDataProvider')
+const TaskTreeItem = require('../views/storyInfo/taskAndBlockerTreeItem')
+const StoryInfoDataProvider = require('../views/storyInfo/storyInfoDataProvider')
 const Context = require('../../test/factories/vscode/context')
 const PtTask = require('../../model/tasks')
