+++ lib/commands/unresolveBlocker.spec.js
@@ -2,8 +2,8 @@ jest.mock('../validation/rebounds')
 jest.mock('../../model/blockers')
 const unresolveBlocker = require('./unresolveBlocker')
 const rebounds = require('../validation/rebounds')
-const TaskTreeItem = require('../views/taskAndBlockerTreeItem')
-const StoryInfoDataProvider = require('../views/storyInfoDataProvider')
+const TaskTreeItem = require('../views/storyInfo/taskAndBlockerTreeItem')
+const StoryInfoDataProvider = require('../views/storyInfo/storyInfoDataProvider')
 const Context = require('../../test/factories/vscode/context')
 const PtBlocker = require('../../model/blockers')
