+++ lib/views/memberCycletime/cycleTimeDataProvider.js
@@ -1,8 +1,8 @@
 const {TreeItem, TreeItemCollapsibleState} = require('vscode')
-const PtIterations = require('../../model/iterations')
-const {commands} = require('../commands')
+const PtIterations = require('../../../model/iterations')
+const {commands} = require('../../commands')
 const IterationPortalItem = require('./iterationPortalItem')
-const {testMatch, expressions} = require('../utils/regularExpressions')
+const {testMatch, expressions} = require('../../utils/regularExpressions')
 
 class CycleTimeTreeDataProvider {
   constructor(context) {
