+++ lib/views/memberCycletime/cycleTimeDataProvider.spec.js
@@ -1,8 +1,8 @@
 jest.mock('../../model/iterations')
 const cycleTimeDataProvider = require('./cycleTimeDataProvider')
-const PtIterations = require('../../model/iterations')
+const PtIterations = require('../../../model/iterations')
 const {TreeItem} = require('vscode')
-const {commands} = require('../commands')
+const {commands} = require('../../commands')
 
 const portal = {
   2018: {
