+++ lib/views/memberCycletime/iterationPortalItem.js
@@ -1,4 +1,4 @@
-const {formatDate} = require('../utils/date')
+const {formatDate} = require('../../utils/date')
 
 
 class IterationPortalItem {
