+++ lib/views/storyInfo/storyInfoDataProvider.js
@@ -1,9 +1,9 @@
 const {TreeItem, TreeItemCollapsibleState, EventEmitter} = require('vscode')
-const {getState} = require('../helpers/pivotaly')
-const PtTask = require('../../model/tasks')
-const PtBlocker = require('../../model/blockers')
+const {getState} = require('../../helpers/pivotaly')
+const PtTask = require('../../../model/tasks')
+const PtBlocker = require('../../../model/blockers')
 const TaskAndBlockerTreeItem = require('./taskAndBlockerTreeItem')
-const commandRepo = require('../commands')
+const commandRepo = require('../../commands')
 
 class StoryInfoDataProvider {
   constructor(context) {
