+++ lib/views/storyInfo/storyInfoDataProvider.spec.js
@@ -1,9 +1,9 @@
 jest.mock('../../model/tasks')
 jest.mock('../../model/blockers')
 const StoryInfoDataProvider = require('./storyInfoDataProvider')
-const Context = require('../../test/factories/vscode/context')
-const PtTasks = require('../../model/tasks')
-const PtBlockers = require('../../model/blockers')
+const Context = require('../../../test/factories/vscode/context')
+const PtTasks = require('../../../model/tasks')
+const PtBlockers = require('../../../model/blockers')
 
 const context = Context.build()
 const StoryInfoProvider = new StoryInfoDataProvider(context)
