+++ lib/views/storyInfo/taskAndBlockerTreeItem.spec.js
@@ -1,6 +1,6 @@
 const TaskAndBlockerTreeItem = require('./taskAndBlockerTreeItem')
 const StoryInfoDataProvider = require('./storyInfoDataProvider')
-const Context = require('../../test/factories/vscode/context')
+const Context = require('../../../test/factories/vscode/context')
 
 const context = Context.build()
 const StoryInfoProvider = new StoryInfoDataProvider(context)
